refactor: 당첨 내역 계산을 재사용할 수 있도록 리팩터링
- 기존에는 하드코딩된 방식 때문에 당첨 내역을 다시 사용하기 어려웠습니다.
- 이제 MatchCondition을 통해 가독성 좋고 확장 가능합니다.
